Create task04.java
This Java program demonstrates basic array handling and string comparison. It creates an array of 6 string variables and allows the user to input 6 names. After the names are entered, the program checks if the name "Ali" is present in the array, ignoring the case of the letters. The program then prints whether "Ali" is found or not in the array. This program is designed to be simple and easy to understand for beginners learning Java.