Create task06.java
This Java program uses the ternary operator to check a user's eligibility for voting based on their age. The program prompts the user to enter their age. It then checks if the user is 18 or older using the ternary operator. If the user's age is 18 or more, the program will display "You are eligible for voting." If the age is less than 18, it will display "You are not eligible for voting." This program demonstrates the use of the ternary operator for conditional checking in a basic and easy-to-understand way.