Create task01.java
This Java program generates prime numbers within a user-defined range. It starts by prompting the user to input the starting and ending numbers of the range. Then, for each number in that range, it checks if the number is prime. To do this, the program checks if the number can be divided by any number smaller than itself (other than 1). If a number is divisible by any number other than 1 and itself, it's not prime. If no such divisors are found, the number is considered prime and is printed. The program continues this process for all numbers within the given range. This simple approach ensures that the user can easily identify prime numbers within any range they choose.